move GKE basic operations to a separate package
